backend/requirements.txt
faiss_cpu==1.12.0
fastapi==0.116.1
numpy==2.2.4
pydantic==2.11.8
python-dotenv==1.1.1
scikit-learn==1.7.2
sentence_transformers==5.1.0
slowapi==0.1.9
uvicorn[standard]==0.35.0
google-generativeai==0.8.5